#87 move vehicle related class to vehicle packages
