+++ src/com/lftechnology/java/training/sanish/generics/vehicle/Car.java
@@ -1,7 +1,8 @@
-package com.lftechnology.java.training.sanish.generics;
+package com.lftechnology.java.training.sanish.generics.vehicle;
 
 /**
  * Car class
+ *
  * @author Sanish Maharjan <sanishmaharjan@lftechnology.com>
  */
 public class Car implements Vehicle {
